# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/blogs": {
      "post": {
        "operationId": "BlogsController_createBlog",
        "summary": "Create a new Blog",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBlogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The Blog has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if title is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/blogs/search": {
      "post": {
        "operationId": "BlogsController_searchAllBlogs",
        "summary": "Search blogs by title",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchBlogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Return all blogs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Blog"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/blogs/sitemap": {
      "get": {
        "operationId": "BlogsController_getBlogSitemap",
        "summary": "Get blogs for sitemap",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return the sitemap.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogSitemapDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/blogs/{id}": {
      "get": {
        "operationId": "BlogsController_getBlogById",
        "summary": "Get a Blog by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the Blog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if Blog is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "BlogsController_updateBlog",
        "summary": "Update a blog with id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlogDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Return the blog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if blog is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "BlogsController_removeBlog",
        "summary": "Remove a blog with id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Return the blog.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Blog"
                }
              }
            }
          },
          "404": {
            "description": "Throws exception if blog is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/blogs/slug/{slug}": {
      "get": {
        "operationId": "BlogsController_getBlogBySlug",
        "summary": "Get a certificate by slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the certificate."
          },
          "404": {
            "description": "Throws exception if certificate is not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundException"
                }
              }
            }
          }
        },
        "tags": [
          "Blogs"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Pomen API",
    "description": "Our API provides a seamless way to create and manage memorial profiles. It allows developers to integrate features for storing and sharing memories, uploading photos and videos, and locating memorial sites. With built-in search functionality, users can easily find profiles by name and explore their biographies, family connections, and important life events. The API supports personalization options, including profile updates and interaction with guestbooks. Designed for scalability and ease of use, our API empowers developers to build meaningful experiences around preserving memories and honoring loved ones.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BlogText": {
        "type": "object",
        "properties": {
          "blogTextId": {
            "type": "number"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "blogContentId": {
            "type": "number"
          },
          "blogContent": {
            "$ref": "#/components/schemas/BlogContent"
          }
        },
        "required": [
          "blogTextId",
          "text",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "blogContentId",
          "blogContent"
        ]
      },
      "Blog": {
        "type": "object",
        "properties": {
          "blogId": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "contents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogContent"
            }
          }
        },
        "required": [
          "blogId",
          "slug",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "contents"
        ]
      },
      "BlogContent": {
        "type": "object",
        "properties": {
          "blogContentId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          },
          "paragraphs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlogText"
            }
          },
          "blogId": {
            "type": "number"
          },
          "blog": {
            "$ref": "#/components/schemas/Blog"
          },
          "type": {
            "type": "object"
          },
          "blogContentImageId": {
            "type": "number"
          },
          "blogContentImage": {
            "$ref": "#/components/schemas/File"
          }
        },
        "required": [
          "blogContentId",
          "createdAt",
          "updatedAt",
          "deletedAt",
          "paragraphs",
          "blogId",
          "blog",
          "type",
          "blogContentImageId",
          "blogContentImage"
        ]
      },
      "File": {
        "type": "object",
        "properties": {}
      },
...
      "NotFoundException": {
        "type": "object",
        "properties": {}
      },
...
      "BlogContentTypeEnum": {
        "type": "string",
        "description": "Type must be one of the following: title, text.center, text.left, text.right",
        "enum": [
          "title",
          "text.center",
          "text.left",
          "text.right"
        ]
      },
      "CreateBlogContentDto": {
        "type": "object",
        "properties": {
          "paragraphs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "$ref": "#/components/schemas/BlogContentTypeEnum"
          }
        },
        "required": [
          "paragraphs",
          "type"
        ]
      },
      "CreateBlogDto": {
        "type": "object",
        "properties": {
          "contents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateBlogContentDto"
            }
          }
        },
        "required": [
          "contents"
        ]
      },
      "SearchBlogDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "page",
          "limit",
          "title"
        ]
      },
      "BlogSitemapDto": {
        "type": "object",
        "properties": {
          "blogId": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "blogId",
          "slug",
          "image",
          "updatedAt"
        ]
      },
      "UpdateBlogTextDto": {
        "type": "object",
        "properties": {
          "blogTextId": {
            "type": "number"
          },
          "text": {
            "type": "string"
          }
        },
        "required": [
          "blogTextId",
          "text"
        ]
      },
      "UpdateBlogContentDto": {
        "type": "object",
        "properties": {
          "blogContentId": {
            "type": "number"
          },
          "paragraphs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateBlogTextDto"
            }
          },
          "type": {
            "$ref": "#/components/schemas/BlogContentTypeEnum"
          }
        },
        "required": [
          "blogContentId",
          "paragraphs",
          "type"
        ]
      },
      "UpdateBlogDto": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/UpdateBlogContentDto"
          }
        },
        "required": [
          "content"
        ]
      }
    }
  }
}